Object selection with merge instead rownumber
In some cases the raw tables are not numbered continously some record
numbers are missing. In this case the matching of the IDs are simpler with merge.

This commit changes parcel (BD and BC object groups) and
borderlines (T_HATARVONAL) identification to merge.

For much clearer code border and borderlines ID objects are different.